chore: split siscan utilsinto classes files
